refactor: optimize NAT module interface with index-based candidate access
- Changed p2p_cand_push_local/remote to return index instead of pointer
- Changed nat_punch parameter from address to index (-1 for batch, >=0 for single)
- Removed nat_find_remote_candidate_index helper function
- Updated all callers across signaling and ICE modules
- Added index bounds checking in nat_punch

+++ src/p2p_crypto_extra.c
@@ -123,7 +123,7 @@ int p2p_base64_encode(const uint8_t *src, size_t slen, char *dst, size_t dlen) {
     }
     
     dst[j] = '\0';  /* NUL 结尾 */
-    return j;
+    return (int)j;
 }
 
 /*
@@ -195,7 +195,7 @@ int p2p_base64_decode(const char *src, size_t slen, uint8_t *dst, size_t dlen) {
         if (src[i - 1] != '=') dst[j++] = (values[2] << 6) | values[3];
     }
     
-    return j;
+    return (int)j;
 }
 
 /* ============================= DES 加密实现 ============================= */
